requirements.txt
requests>=2.30
nltk==3.9.3
pandas>=2.2,<3
wordcloud==1.9.6
ruff==0.15.6
spacy==3.8.11
scikit-learn==1.8.0
fpdf2==2.8.7
streamlit==1.55.0